third_party/blink/web_tests/external/wpt/mathml/presentation-markup/operators/mo-stretch-properties-dynamic-001-expected.txt

This is a testharness.js-based test.
FAIL minsize assert_approx_equals: attach expected 100 +/- 1 but got 25
FAIL maxsize assert_approx_equals: attach expected 100 +/- 1 but got 25
FAIL largeop assert_approx_equals: set true expected 50 +/- 1 but got 25
FAIL symmetric assert_approx_equals: set true expected 150 +/- 1 but got 25
FAIL stretchy assert_approx_equals: set true expected 75 +/- 1 but got 25
Harness: the test ran to completion.
